+++ pybamm/models/submodels/interface/lithium_plating/no_plating.py
@@ -14,12 +14,14 @@ class NoPlating(BasePlating):
         The parameters to use for this submodel
     domain : str
         The domain of the model either 'Negative' or 'Positive'
+    options : dict, optional
+        A dictionary of options to be passed to the model.
 
     **Extends:** :class:`pybamm.lithium_plating.BasePlating`
     """
 
-    def __init__(self, param, domain):
-        super().__init__(param, domain)
+    def __init__(self, param, domain, options=None):
+        super().__init__(param, domain, options=options)
 
     def get_fundamental_variables(self):
         zero = pybamm.FullBroadcast(