+++ tests/integration/test_models/test_full_battery_models/test_lithium_ion/test_basic_half_cell_models.py
@@ -19,16 +19,6 @@ class TestBasicHalfCellModels(unittest.TestCase):
         chemistry = pybamm.parameter_sets.Xu2019
         param = pybamm.ParameterValues(chemistry=chemistry)
 
-        # add lithium counter electrode parameter values
-        param.update(
-            {
-                "Lithium counter electrode exchange-current density [A.m-2]": 12.6,
-                "Lithium counter electrode conductivity [S.m-1]": 1.0776e7,
-                "Lithium counter electrode thickness [m]": 250e-6,
-            },
-            check_already_exists=False,
-        )
-
         param["Current function [A]"] = 2.5e-3
 
         # process model and geometry
@@ -57,20 +47,9 @@ class TestBasicHalfCellModels(unittest.TestCase):
         geometry = model.default_geometry
 
         # load parameter values
-        chemistry = pybamm.parameter_sets.Chen2020
+        chemistry = pybamm.parameter_sets.Chen2020_plating
         param = pybamm.ParameterValues(chemistry=chemistry)
 
-        # add lithium counter electrode parameter values
-        param.update(
-            {
-                "Lithium counter electrode exchange-current density [A.m-2]": 12.6,
-                "Lithium counter electrode conductivity [S.m-1]": 1.0776e7,
-                "Lithium counter electrode thickness [m]": 250e-6,
-            },
-            check_already_exists=False,
-        )
-
-        param["Initial concentration in negative electrode [mol.m-3]"] = 1000
         param["Current function [A]"] = 2.5
 
         # process model and geometry